+++ eslint.config.mjs
@@ -1,0 +1,8 @@
+import baseConfig from './packages/eslint-config/index.js'
+
+export default [
+  ...baseConfig,
+  {
+    ignores: ['**/dist/*'],
+  },
+]