+++ packages/ajv-validator/test/index.test.ts
@@ -1,12 +1,12 @@
-import { Hono } from 'hono';
-import type { Equal, Expect } from 'hono/utils/types';
-import { ajvValidator } from '../src';
-import { JSONSchemaType, type ErrorObject } from 'ajv';
+import type { JSONSchemaType, type ErrorObject } from 'ajv'
+import { Hono } from 'hono'
+import type { Equal, Expect } from 'hono/utils/types'
+import { ajvValidator } from '../src'
 
-type ExtractSchema<T> = T extends Hono<infer _, infer S> ? S : never;
+type ExtractSchema<T> = T extends Hono<infer _, infer S> ? S : never
 
 describe('Basic', () => {
-  const app = new Hono();
+  const app = new Hono()
 
   const schema: JSONSchemaType<{ name: string; age: number }> = {
     type: 'object',
@@ -16,35 +16,35 @@ describe('Basic', () => {
     },
     required: ['name', 'age'],
     additionalProperties: false,
-  };
+  }
 
   const route = app.post('/author', ajvValidator('json', schema), (c) => {
-    const data = c.req.valid('json');
+    const data = c.req.valid('json')
     return c.json({
       success: true,
       message: `${data.name} is ${data.age}`,
-    });
-  });
+    })
+  })
 
-  type Actual = ExtractSchema<typeof route>;
+  type Actual = ExtractSchema<typeof route>
   type Expected = {
     '/author': {
       $post: {
         input: {
           json: {
-            name: string;
-            age: number;
-          };
-        };
+            name: string
+            age: number
+          }
+        }
         output: {
-          success: boolean;
-          message: string;
-        };
-      };
-    };
-  };
+          success: boolean
+          message: string
+        }
+      }
+    }
+  }
 
-  type verify = Expect<Equal<Expected, Actual>>;
+  type verify = Expect<Equal<Expected, Actual>>
 
   it('Should return 200 response', async () => {
     const req = new Request('http://localhost/author', {
@@ -56,15 +56,15 @@ describe('Basic', () => {
       headers: {
         'Content-Type': 'application/json',
       },
-    });
-    const res = await app.request(req);
-    expect(res).not.toBeNull();
-    expect(res.status).toBe(200);
+    })
+    const res = await app.request(req)
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
     expect(await res.json()).toEqual({
       success: true,
       message: 'Superman is 20',
-    });
-  });
+    })
+  })
 
   it('Should return 400 response', async () => {
     const req = new Request('http://localhost/author', {
@@ -76,17 +76,17 @@ describe('Basic', () => {
       headers: {
         'Content-Type': 'application/json',
       },
-    });
-    const res = await app.request(req);
-    expect(res).not.toBeNull();
-    expect(res.status).toBe(400);
-    const data = (await res.json()) as { success: boolean };
-    expect(data.success).toBe(false);
-  });
-});
+    })
+    const res = await app.request(req)
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(400)
+    const data = (await res.json()) as { success: boolean }
+    expect(data.success).toBe(false)
+  })
+})
 
 describe('With Hook', () => {
-  const app = new Hono();
+  const app = new Hono()
 
   const schema: JSONSchemaType<{ id: number; title: string }> = {
     type: 'object',
@@ -96,39 +96,39 @@ describe('With Hook', () => {
     },
     required: ['id', 'title'],
     additionalProperties: false,
-  };
+  }
 
   app
     .post(
       '/post',
       ajvValidator('json', schema, (result, c) => {
         if (!result.success) {
-          return c.text('Invalid!', 400);
+          return c.text('Invalid!', 400)
         }
-        const data = result.data;
-        return c.text(`${data.id} is valid!`);
+        const data = result.data
+        return c.text(`${data.id} is valid!`)
       }),
       (c) => {
-        const data = c.req.valid('json');
+        const data = c.req.valid('json')
         return c.json({
           success: true,
           message: `${data.id} is ${data.title}`,
-        });
+        })
       }
     )
     .post(
       '/errorTest',
       ajvValidator('json', schema, (result, c) => {
-        return c.json(result, 400);
+        return c.json(result, 400)
       }),
       (c) => {
-        const data = c.req.valid('json');
+        const data = c.req.valid('json')
         return c.json({
           success: true,
           message: `${data.id} is ${data.title}`,
-        });
+        })
       }
-    );
+    )
 
   it('Should return 200 response', async () => {
     const req = new Request('http://localhost/post', {
@@ -140,12 +140,12 @@ describe('With Hook', () => {
       headers: {
         'Content-Type': 'application/json',
       },
-    });
-    const res = await app.request(req);
-    expect(res).not.toBeNull();
-    expect(res.status).toBe(200);
-    expect(await res.text()).toBe('123 is valid!');
-  });
+    })
+    const res = await app.request(req)
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
+    expect(await res.text()).toBe('123 is valid!')
+  })
 
   it('Should return 400 response', async () => {
     const req = new Request('http://localhost/post', {
@@ -157,11 +157,11 @@ describe('With Hook', () => {
       headers: {
         'Content-Type': 'application/json',
       },
-    });
-    const res = await app.request(req);
-    expect(res).not.toBeNull();
-    expect(res.status).toBe(400);
-  });
+    })
+    const res = await app.request(req)
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(400)
+  })
 
   it('Should return 400 response and error array', async () => {
     const req = new Request('http://localhost/errorTest', {
@@ -172,17 +172,17 @@ describe('With Hook', () => {
       headers: {
         'Content-Type': 'application/json',
       },
-    });
-    const res = await app.request(req);
-    expect(res).not.toBeNull();
-    expect(res.status).toBe(400);
+    })
+    const res = await app.request(req)
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(400)
 
     const { errors, success } = (await res.json()) as {
-      success: boolean;
-      errors: ErrorObject[];
-    };
-    expect(success).toBe(false);
-    expect(Array.isArray(errors)).toBe(true);
+      success: boolean
+      errors: ErrorObject[]
+    }
+    expect(success).toBe(false)
+    expect(Array.isArray(errors)).toBe(true)
     expect(
       errors.map((e: ErrorObject) => ({
         keyword: e.keyword,
@@ -195,6 +195,6 @@ describe('With Hook', () => {
         instancePath: '',
         message: "must have required property 'title'",
       },
-    ]);
-  });
-});
+    ])
+  })
+})