+++ packages/conform-validator/test/valibot.test.ts
@@ -1,10 +1,10 @@
-import type { ExtractSchema, ParsedFormValue } from 'hono/types'
-import type { Equal, Expect } from 'hono/utils/types'
-import type { StatusCode } from 'hono/utils/http-status'
-import * as v from 'valibot'
+import { parseWithValibot } from 'conform-to-valibot'
 import { Hono } from 'hono'
 import { hc } from 'hono/client'
-import { parseWithValibot } from 'conform-to-valibot'
+import type { ExtractSchema, ParsedFormValue } from 'hono/types'
+import type { StatusCode } from 'hono/utils/http-status'
+import type { Equal, Expect } from 'hono/utils/types'
+import * as v from 'valibot'
 import { conformValidator } from '../src'
 
 describe('Validate requests using a Valibot schema', () => {